requirements.txt
beautifulsoup4
langchain
langchain-chroma
langchain-community
langchain-core
langchain-google-genai
langchain-openai
langchain-text-splitters
selenium
streamlit
webdriver-manager
python-dotenv
langchain-hub
jq
pprint